+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Enhancements
 
-* Support `dif` and `lex` in formula generation
+* Support `dif`, `dir` and `lex` in formula generation
 * Support `bin` (binomial coefficient) in LEAN formula generation
 
 ## v26.1.25